Remove old HieDB indexer parts
