--- glean/lang/codemarkup/tests/haskell/cases/input_key/entity_location.out
@@ -1,47 +1,0 @@
-[
-  "@generated",
-  {
-    "key": {
-      "entity": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "length": 1, "start": 281 }
-              }
-            }
-          }
-        }
-      },
-      "location": {
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-        "location": { "span": { "length": 1, "start": 281 } },
-        "name": "A.a"
-      }
-    }
-  },
-  {
-    "key": {
-      "entity": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "B.b" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-                "span": { "length": 1, "start": 284 }
-              }
-            }
-          }
-        }
-      },
-      "location": {
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-        "location": { "span": { "length": 1, "start": 284 } },
-        "name": "B.b"
-      }
-    }
-  }
-]
