--- glean/lang/codemarkup/tests/haskell/cases/input_key/entity_location.query
@@ -1,2 +1,0 @@
-query: codemarkup.EntityLocation _
-transform: [normord, []]
