--- glean/lang/codemarkup/tests/haskell/cases/input_key/entity_uses.out
@@ -1,136 +1,0 @@
-[
-  "@generated",
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "start": 281, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-      "span": { "start": 228, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "start": 281, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-      "span": { "start": 269, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "start": 281, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-      "span": { "start": 281, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "start": 281, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-      "span": { "start": 306, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "B.b" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-                "span": { "start": 284, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-      "span": { "start": 228, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "B.b" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-                "span": { "start": 284, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-      "span": { "start": 273, "length": 1 }
-    }
-  },
-  {
-    "key": {
-      "target": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "B.b" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-                "span": { "start": 284, "length": 1 }
-              }
-            }
-          }
-        }
-      },
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-      "span": { "start": 284, "length": 1 }
-    }
-  }
-]
