--- glean/lang/codemarkup/tests/haskell/cases/input_key/entity_uses.perf
@@ -1,5 +1,0 @@
-{
-  "@generated": null,
-  "facts_searched": { "hs.TargetUses.2": 3 },
-  "full_scans": [ "hs.TargetUses.2" ]
-}
