--- glean/lang/codemarkup/tests/haskell/cases/input_key/entity_uses.query
@@ -1,3 +1,0 @@
-query: codemarkup.EntityUses _
-perf: true
-transform: [gensort, []]
