--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_locations.out
@@ -1,49 +1,0 @@
-[
-  "@generated",
-  {
-    "key": {
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-      "location": {
-        "name": "A.a",
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-        "location": { "span": { "start": 281, "length": 1 } }
-      },
-      "entity": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "A.a" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-                "span": { "start": 281, "length": 1 }
-              }
-            }
-          }
-        }
-      }
-    }
-  },
-  {
-    "key": {
-      "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-      "location": {
-        "name": "B.b",
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-        "location": { "span": { "start": 284, "length": 1 } }
-      },
-      "entity": {
-        "hs": {
-          "definition": {
-            "key": {
-              "name": { "key": "B.b" },
-              "source": {
-                "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-                "span": { "start": 284, "length": 1 }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-]
