--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_locations.perf
@@ -1,5 +1,0 @@
-{
-  "@generated": null,
-  "facts_searched": { "hs.FileDefinition.2": 2 },
-  "full_scans": [ "hs.FileDefinition.2" ]
-}
