--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_locations.query
@@ -1,3 +1,0 @@
-query: codemarkup.FileEntityLocations _
-perf: true
-transform: [gensort, []]
