--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_xrefs.perf
@@ -1,9 +1,0 @@
-{
-  "@generated": null,
-  "facts_searched": {
-    "hs.Definition.2": 7,
-    "hs.DefinitionName.1": 12,
-    "hs.FileXRefMap.2": 2
-  },
-  "full_scans": [ "hs.FileXRefMap.2" ]
-}
