--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_xrefs.query
@@ -1,3 +1,0 @@
-query: codemarkup.FileEntityXRefLocations _
-perf: true
-transform: [gensort, []]
