--- glean/lang/codemarkup/tests/haskell/cases/input_key/file_entity_xrefs_generic.query
@@ -1,3 +1,0 @@
-query: codemarkup.FileXRefsGenericEntities _
-perf: true
-transform: [gensort, []]
