--- glean/lang/codemarkup/tests/haskell/cases/input_key/resolve_location.perf
@@ -1,5 +1,0 @@
-{
-  "@generated": null,
-  "facts_searched": { "hs.FileDefinition.2": 2 },
-  "full_scans": [ "hs.FileDefinition.2" ]
-}
