--- glean/lang/codemarkup/tests/haskell/cases/input_key/resolve_location.query
@@ -1,3 +1,0 @@
-query: codemarkup.ResolveLocation _
-perf: true
-transform: [gensort, []]
