--- glean/lang/haskell/tests/cases/library/hsDefinition.out
@@ -1,21 +1,0 @@
-[
-  "@generated",
-  {
-    "key": {
-      "name": { "key": "A.a" },
-      "source": {
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-        "span": { "start": 281, "length": 1 }
-      }
-    }
-  },
-  {
-    "key": {
-      "name": { "key": "B.b" },
-      "source": {
-        "file": { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" },
-        "span": { "start": 284, "length": 1 }
-      }
-    }
-  }
-]
