--- glean/lang/haskell/tests/cases/library/hsDefinition.query
@@ -1,2 +1,0 @@
-query: hs.Definition _
-transform: [gensort, []]
