--- glean/lang/haskell/tests/cases/library/hsDefinitionName.out
@@ -1,13 +1,0 @@
-[
-  "@generated",
-  { "key": "A.a" },
-  { "key": "B.b" },
-  { "key": "Data.String.IsString[]" },
-  { "key": "GHC.Base.$" },
-  { "key": "GHC.Base.String" },
-  { "key": "GHC.Base.map" },
-  { "key": "GHC.Base.ord" },
-  { "key": "GHC.List.reverse" },
-  { "key": "GHC.Types.Int" },
-  { "key": "GHC.Unicode.toLower" }
-]
