--- glean/lang/haskell/tests/cases/library/hsDefinitionName.query
@@ -1,2 +1,0 @@
-query: hs.DefinitionName _
-transform: [gensort, []]
