--- glean/lang/haskell/tests/cases/library/hsFileDefinition.query
@@ -1,2 +1,0 @@
-query: hs.FileDefinition _
-transform: [gensort, []]
