--- glean/lang/haskell/tests/cases/library/hsFileXRefMap.query
@@ -1,2 +1,0 @@
-query: hs.FileXRefMap _
-transform: [gensort, []]
