--- glean/lang/haskell/tests/cases/library/hsTargetUses.query
@@ -1,2 +1,0 @@
-query: hs.TargetUses _
-transform: [gensort, []]
