--- glean/lang/haskell/tests/cases/library/srcfile.out
@@ -1,5 +1,0 @@
-[
-  "@generated",
-  { "key": "fbcode/glean/lang/haskell/tests/code/A.hs" },
-  { "key": "fbcode/glean/lang/haskell/tests/code/B.hs" }
-]
