--- glean/lang/haskell/tests/cases/library/srcfile.query
@@ -1,2 +1,0 @@
-query: src.File _
-transform: [gensort, []]
